python_project/videoproject/requirement.txt
djangorestframework
moviepy
celery
redis
flower
django-cors-headers
gunicorn
django-simpleui
feedparser
requests
beautifulsoup4
lxml
djangorestframework-simplejwt